allowing users to modify goals with confirmation
