+++ pkg/commands/modify/actions/errors.go
@@ -1,0 +1,7 @@
+package actions
+
+type NoModificationError struct{}
+
+func (m *NoModificationError) Error() string {
+	return "no values provided for modification."
+}
