+++ pkg/commands/modify/modify.go
@@ -1,7 +1,7 @@
 package modify
 
 import (
-	"github.com/sqljames/goalctl/pkg/flags"
+	"github.com/sqljames/goalctl/pkg/commands/modify/actions"
 	"github.com/urfave/cli/v2"
 )
 
@@ -11,47 +11,7 @@ func New() *cli.Command {
 		Usage:   "Allows users to modify objects",
 		Aliases: []string{"m"},
 		Subcommands: []*cli.Command{
-			modifyGoal(),
+			actions.ModifyGoal(),
 		},
 	}
-}
-
-var (
-	NameGoalFlag *cli.StringFlag = &cli.StringFlag{
-		Name:     flags.NameFlagName,
-		Usage:    "Name of the Goal.",
-		Required: false,
-		//Category: "formatting",
-		//Aliases: []string{"nb"},
-	}
-	GoalIDFlag *cli.StringFlag = &cli.StringFlag{
-		Name:     flags.GoalIDFlagName,
-		Usage:    "ID of the Goal you want to link with the log entry",
-		Required: true,
-		Aliases:  []string{"g"},
-	}
-
-	EntryFlag *cli.StringFlag = &cli.StringFlag{
-		Name:     flags.EntryTextFlagName,
-		Usage:    "Details of your object",
-		Required: false,
-		//Category: "formatting",
-		//Aliases: []string{"nb"},
-	}
-)
-
-func modifyGoal() *cli.Command {
-	return &cli.Command{
-		Name:   "goal",
-		Usage:  "Allows users to modify a target goal",
-		Action: actionModifyGoal,
-		Flags: []cli.Flag{
-			GoalIDFlag,
-			flags.PriorityFlag,
-			flags.GoalStatusFlag,
-			flags.DueDateFlag,
-			EntryFlag,
-			NameGoalFlag,
-		},
-	} 
 }
